e2e: Improve e2e UI test speed by condensing tests into one file
Change-type: patch
